Removed ListenableJobBoard content into JobBoard
In other words, massively simplified the boards' code.
