--- src/main/java/dte/employme/board/ForwardingJobBoard.java
@@ -1,62 +1,0 @@
-package dte.employme.board;
-
-import java.util.Iterator;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-import org.bukkit.entity.Player;
-
-import dte.employme.job.Job;
-
-public abstract class ForwardingJobBoard implements JobBoard
-{
-	protected final JobBoard delegate;
-	
-	protected ForwardingJobBoard(JobBoard delegate) 
-	{
-		this.delegate = delegate;
-	}
-	
-	@Override
-	public void addJob(Job job)
-	{
-		this.delegate.addJob(job);
-	}
-
-	@Override
-	public void removeJob(Job job)
-	{
-		this.delegate.removeJob(job);
-	}
-
-	@Override
-	public void completeJob(Job job, Player whoCompleted) 
-	{
-		this.delegate.completeJob(job, whoCompleted);
-	}
-
-	@Override
-	public List<Job> getOfferedJobs() 
-	{
-		return this.delegate.getOfferedJobs();
-	}
-	
-	@Override
-	public Optional<Job> getJobByUUID(UUID uuid) 
-	{
-		return this.delegate.getJobByUUID(uuid);
-	}
-
-	@Override
-	public List<Job> getJobsOfferedBy(UUID employerUUID) 
-	{
-		return this.delegate.getJobsOfferedBy(employerUUID);
-	}
-	
-	@Override
-	public Iterator<Job> iterator() 
-	{
-		return this.delegate.iterator();
-	}
-}
